Beginnings of an arrowAnnotate conversion
